add test cases for ts-4454

+++ tests/script/sh/stop_dnodes.sh
@@ -1,56 +1,37 @@
 #!/bin/sh
-
+# shellcheck disable=SC2009
+# shellcheck disable=SC2086
 set +e
 #set -x
 
 unset LD_PRELOAD
-UNAME_BIN=`which uname`
-OS_TYPE=`$UNAME_BIN`
+UNAME_BIN=$(which uname)
+OS_TYPE=$($UNAME_BIN)
 
-PID=`ps -ef|grep /usr/bin/taosd | grep -v grep | awk '{print $2}'`
-if [ -n "$PID" ]; then
+kill_process() {
+    PID=$(ps -ef|grep -w "$1" | grep -v grep | grep -v defunct |  awk '{print $2}')
+    while [ -n "$PID" ]; do
+      # shellcheck disable=SC2086
+      echo "killing $1 processes, pid: $PID"
+      kill -9 $PID
+      if [ "$1" = "taosd" ]; then
+        if [ "$OS_TYPE" != "Darwin" ]; then
+          fuser -k -n tcp 6030
+        else
+          lsof -nti:6030 | xargs kill -9
+        fi
+      fi
+      PID=$(ps -ef|grep "$1" | grep -v grep | grep -v defunct |  awk '{print $2}')
+    done
+}
+
+PID1=$(ps -ef|grep /usr/bin/taosd | grep -v grep | grep -v defunct |  awk '{print $2}')
+if [ -n "$PID1" ]; then
   echo systemctl stop taosd
   systemctl stop taosd
 fi
 
-PID=`ps -ef|grep -w taosd | grep -v grep | awk '{print $2}'`
-while [ -n "$PID" ]; do
-  echo kill -9 $PID
-  #pkill -9 taosd
-  kill -9 $PID
-  echo "Killing taosd processes"
-  if [ "$OS_TYPE" != "Darwin" ]; then
-    fuser -k -n tcp 6030
-  else
-    lsof -nti:6030 | xargs kill -9
-  fi
-  PID=`ps -ef|grep -w taosd | grep -v grep | awk '{print $2}'`
-done
-
-PID=`ps -ef|grep -w taos | grep -v grep | awk '{print $2}'`
-while [ -n "$PID" ]; do
-  echo kill -9 $PID
-  #pkill -9 taos
-  kill -9 $PID
-  echo "Killing taos processes"
-  if [ "$OS_TYPE" != "Darwin" ]; then
-    fuser -k -n tcp 6030
-  else
-    lsof -nti:6030 | xargs kill -9
-  fi
-  PID=`ps -ef|grep -w taos | grep -v grep | awk '{print $2}'`
-done
-
-PID=`ps -ef|grep -w tmq_sim | grep -v grep | awk '{print $2}'`
-while [ -n "$PID" ]; do
-  echo kill -9 $PID
-  #pkill -9 tmq_sim
-  kill -9 $PID
-  echo "Killing tmq_sim processes"
-  if [ "$OS_TYPE" != "Darwin" ]; then
-    fuser -k -n tcp 6030
-  else
-    lsof -nti:6030 | xargs kill -9
-  fi
-  PID=`ps -ef|grep -w tmq_sim | grep -v grep | awk '{print $2}'`
-done
+kill_process taosd
+kill_process taos
+kill_process taosadapter
+kill_process tmq_sim
